requirements.txt
cmk_agent_based
cmk_mkp_tool
cmk_rulesets

isort
wemake-python-styleguide